backend/requirements.txt
fastapi
starlette
pydantic
supabase
passlib
authlib
python-dotenv
tensorflow
numpy
google-generativeai
jinja2